+++ apps/web/test/test-utils/test-utils.ts
@@ -70,7 +70,7 @@ export function stubClient(): MatrixClient {
     vi.spyOn(peg, "get");
     vi.spyOn(peg, "safeGet");
     vi.spyOn(peg, "unset");
-    vi.spyOn(peg, "replaceUsingCreds");
+    vi.spyOn(peg, "set");
     // MatrixClientPeg.safeGet() is called a /lot/, so implement it with our own
     // fast stub function rather than a sinon stub
     peg.get = () => client;
