+++ apps/web/test/unit-tests/Lifecycle-test.ts
@@ -29,6 +29,7 @@ import { persistAccessTokenInStorage, persistRefreshTokenInStorage } from "../..
 import { encryptPickleKey } from "../../src/utils/tokens/pickling";
 import * as StorageManager from "../../src/utils/StorageManager.ts";
 import type BasePlatform from "../../src/BasePlatform.ts";
+import * as createMatrixClientModule from "../../src/utils/createMatrixClient";
 
 const { logout, restoreSessionFromStorage, setLoggedIn } = Lifecycle;
 
@@ -71,9 +72,11 @@ describe("Lifecycle", () => {
             logout: jest.fn().mockResolvedValue(undefined),
             getAccessToken: jest.fn(),
             getRefreshToken: jest.fn(),
+            setGuest: jest.fn(),
+            setNotifTimelineSet: jest.fn(),
         });
         // stub this
-        jest.spyOn(MatrixClientPeg, "replaceUsingCreds").mockImplementation(() => {});
+        jest.spyOn(MatrixClientPeg, "set").mockImplementation(() => {});
         jest.spyOn(MatrixClientPeg, "start").mockResolvedValue(undefined);
 
         // reset any mocking
@@ -187,6 +190,7 @@ describe("Lifecycle", () => {
             jest.spyOn(logger, "log").mockClear();
 
             jest.spyOn(MatrixJs, "createClient").mockReturnValue(mockClient);
+            jest.spyOn(createMatrixClientModule, "createClientWithCreds").mockReturnValue(mockClient);
 
             // stub this out
             jest.spyOn(Modal, "createDialog").mockReturnValue(
@@ -235,7 +239,7 @@ describe("Lifecycle", () => {
                 it("should restore guest accounts when ignoreGuest is false", async () => {
                     expect(await restoreSessionFromStorage({ ignoreGuest: false })).toEqual(true);
 
-                    expect(MatrixClientPeg.replaceUsingCreds).toHaveBeenCalledWith(
+                    expect(createMatrixClientModule.createClientWithCreds).toHaveBeenCalledWith(
                         expect.objectContaining({
                             userId,
                             guest: true,
@@ -279,7 +283,7 @@ describe("Lifecycle", () => {
                 it("should create and start new matrix client with credentials", async () => {
                     expect(await restoreSessionFromStorage()).toEqual(true);
 
-                    expect(MatrixClientPeg.replaceUsingCreds).toHaveBeenCalledWith(
+                    expect(createMatrixClientModule.createClientWithCreds).toHaveBeenCalledWith(
                         {
                             userId,
                             accessToken,
@@ -328,7 +332,7 @@ describe("Lifecycle", () => {
                     it("should create new matrix client with credentials", async () => {
                         expect(await restoreSessionFromStorage()).toEqual(true);
 
-                        expect(MatrixClientPeg.replaceUsingCreds).toHaveBeenCalledWith(
+                        expect(createMatrixClientModule.createClientWithCreds).toHaveBeenCalledWith(
                             {
                                 userId,
                                 accessToken,
@@ -410,7 +414,7 @@ describe("Lifecycle", () => {
                     expect(await restoreSessionFromStorage()).toEqual(true);
 
                     // Ensure that the expected calls were made
-                    expect(MatrixClientPeg.replaceUsingCreds).toHaveBeenCalledWith(
+                    expect(createMatrixClientModule.createClientWithCreds).toHaveBeenCalledWith(
                         {
                             userId,
                             // decrypted accessToken
@@ -448,7 +452,7 @@ describe("Lifecycle", () => {
                     it("should create new matrix client with credentials", async () => {
                         expect(await restoreSessionFromStorage()).toEqual(true);
 
-                        expect(MatrixClientPeg.replaceUsingCreds).toHaveBeenCalledWith(
+                        expect(createMatrixClientModule.createClientWithCreds).toHaveBeenCalledWith(
                             {
                                 userId,
                                 accessToken,
@@ -501,7 +505,7 @@ describe("Lifecycle", () => {
                     expect(await restoreSessionFromStorage()).toEqual(true);
 
                     // Ensure that the expected calls were made
-                    expect(MatrixClientPeg.replaceUsingCreds).toHaveBeenCalledWith(
+                    expect(createMatrixClientModule.createClientWithCreds).toHaveBeenCalledWith(
                         {
                             userId,
                             // decrypted accessToken
@@ -614,6 +618,7 @@ describe("Lifecycle", () => {
         describe("without a pickle key", () => {
             beforeEach(() => {
                 jest.spyOn(mockPlatform, "createPickleKey").mockResolvedValue(null);
+                jest.spyOn(createMatrixClientModule, "createClientWithCreds").mockReturnValue(mockClient);
             });
 
             it("should persist credentials", async () => {
@@ -667,7 +672,7 @@ describe("Lifecycle", () => {
             it("should create new matrix client with credentials", async () => {
                 expect(await setLoggedIn(credentials)).toEqual(mockClient);
 
-                expect(MatrixClientPeg.replaceUsingCreds).toHaveBeenCalledWith(
+                expect(createMatrixClientModule.createClientWithCreds).toHaveBeenCalledWith(
                     {
                         userId,
                         accessToken,
@@ -763,9 +768,10 @@ describe("Lifecycle", () => {
             });
 
             it("should create new matrix client with credentials", async () => {
+                jest.spyOn(createMatrixClientModule, "createClientWithCreds").mockReturnValue(mockClient);
                 expect(await setLoggedIn(credentials)).toEqual(mockClient);
 
-                expect(MatrixClientPeg.replaceUsingCreds).toHaveBeenCalledWith(
+                expect(createMatrixClientModule.createClientWithCreds).toHaveBeenCalledWith(
                     {
                         userId,
                         accessToken,
@@ -856,7 +862,7 @@ describe("Lifecycle", () => {
                     }),
                 ).toEqual(mockClient);
 
-                expect(MatrixClientPeg.replaceUsingCreds).toHaveBeenCalledWith(
+                expect(createMatrixClientModule.createClientWithCreds).toHaveBeenCalledWith(
                     expect.objectContaining({
                         accessToken,
                         refreshToken,
@@ -877,7 +883,7 @@ describe("Lifecycle", () => {
                     }),
                 ).toEqual(mockClient);
 
-                expect(MatrixClientPeg.replaceUsingCreds).toHaveBeenCalledWith(
+                expect(createMatrixClientModule.createClientWithCreds).toHaveBeenCalledWith(
                     expect.objectContaining({
                         accessToken,
                         refreshToken,
@@ -939,6 +945,7 @@ describe("Lifecycle", () => {
 
         it("should replace the current login with a new one", async () => {
             const stopSpy = jest.spyOn(mockClient, "stopClient").mockReturnValue(undefined);
+            jest.spyOn(createMatrixClientModule, "createClientWithCreds").mockReturnValue(mockClient);
             const dis = window.mxDispatcher;
 
             const firstLoginEvent: Promise<void> = new Promise((resolve) => {
@@ -958,7 +965,7 @@ describe("Lifecycle", () => {
             // So spy on it and make sure it's not called.
             jest.spyOn(MatrixClientPeg, "unset").mockReturnValue(undefined);
 
-            expect(MatrixClientPeg.replaceUsingCreds).toHaveBeenCalledWith(
+            expect(createMatrixClientModule.createClientWithCreds).toHaveBeenCalledWith(
                 expect.objectContaining({
                     userId,
                 }),
@@ -992,7 +999,7 @@ describe("Lifecycle", () => {
             // the client should have been stopped
             expect(stopSpy).toHaveBeenCalledTimes(2);
 
-            expect(MatrixClientPeg.replaceUsingCreds).toHaveBeenCalledWith(
+            expect(createMatrixClientModule.createClientWithCreds).toHaveBeenCalledWith(
                 expect.objectContaining({
                     userId: otherCredentials.userId,
                 }),
