+++ apps/web/test/unit-tests/MatrixClientPeg-test.ts
@@ -7,9 +7,10 @@ Please see LICENSE files in the repository root for full details.
 */
 
 import { logger } from "matrix-js-sdk/src/logger";
+import { type MatrixClient } from "matrix-js-sdk/src/matrix";
 import fetchMock from "@fetch-mock/jest";
 
-import { advanceDateAndTime, stubClient } from "../test-utils";
+import { advanceDateAndTime, createTestClient, stubClient } from "../test-utils";
 import { type IMatrixClientPeg, MatrixClientPeg as peg } from "../../src/MatrixClientPeg";
 
 jest.useFakeTimers();
@@ -70,12 +71,11 @@ describe("MatrixClientPeg", () => {
             // instantiate a MatrixClientPegClass instance, with a new MatrixClient
             testPeg = new PegClass();
             fetchMock.get("http://example.com/_matrix/client/versions", {});
-            testPeg.replaceUsingCreds({
-                accessToken: "SEKRET",
-                homeserverUrl: "http://example.com",
-                userId: "@user:example.com",
-                deviceId: "TEST_DEVICE_ID",
-            });
+
+            const mockClient = createTestClient();
+            mockClient.initRustCrypto = jest.fn();
+            mockClient.startClient = jest.fn();
+            testPeg.set(mockClient as unknown as MatrixClient);
         });
 
         it("should initialise the rust crypto library by default", async () => {
